+++ app/Http/Controllers/SubcategoryController.php
@@ -14,7 +14,7 @@ class SubcategoryController extends Controller
     private CategoryRepository $categoryRepository;
     private SubcategoryRepository $subcategoryRepository;
     private PageRepository $pageRepository;
-    private $type;
+    private string $type;
 
     public function __construct(CategoryRepository $categoryRepository)
     {
@@ -28,6 +28,7 @@ class SubcategoryController extends Controller
     public function show($view, $id)
     {
         $subcategory = $this->subcategoryRepository->getModel()->find($id);
+        $this->authorize('author', $subcategory);
 
         if ($view == 'visible') $pages = $this->pageRepository->getAllByParameters($id, $this->type, 0);
         else if ($view == "hidden") $pages = $this->pageRepository->getAllByParameters($id, $this->type, 1);
@@ -43,10 +44,9 @@ class SubcategoryController extends Controller
     public function showPublic($id)
     {
         $subcategory = $this->subcategoryRepository->getModel()->find($id);
+        $this->authorize('author', $subcategory);
 
-        if ($subcategory->public == 0) {
-            return abort(403, 'Zasób nie jest publiczny');
-        }
+        if ($subcategory->public == 0) return abort(403, 'Zasób nie jest publiczny');
 
         $category = $this->categoryRepository->getModel()->find($subcategory->category_id);
 
@@ -62,20 +62,21 @@ class SubcategoryController extends Controller
     //! CREATE
     public function create(Request $request, $category_id)
     {
-        $category_image = $this->categoryRepository->getModel()->find($category_id)->image_url;
-
+        $category = $this->categoryRepository->getModel()->find($category_id);
+        $this->authorize('categoryAuthor', [new Subcategory(), $category]);
 
         return view('subcategory.create', [
             'category_id' => $category_id,
             'view' => $request->input('view'),
-            'category_image' => $category_image,
+            'category_image' => $category->image_url,
         ]);
     }
 
     public function store(Store $request)
     {
         $data = $request->validated();
-        // $this->authorize('create', [new Subcategory(), $id]);
+        $category = $this->categoryRepository->getModel()->find($data['category_id']);
+        $this->authorize('categoryAuthor', [new Subcategory(), $category]);
 
         if ($request->input('public') != NULL) $data['public'] = true;
         else $data['public'] = false;
@@ -89,30 +90,27 @@ class SubcategoryController extends Controller
     //! EDIT
     public function edit(Request $request, $id)
     {
-        // $this->authorize('edit', [new Subcategory, $id]);
         $subcategory = $this->subcategoryRepository->getModel()->find($id);
-        $category_image = $this->categoryRepository->getModel()->find($subcategory->category_id)->image_url;
+        $this->authorize('author', $subcategory);
+
         return view(
             'subcategory.edit',
             [
                 'subcategory' => $subcategory,
                 'view' => $request->input('view'),
-                'category_image' => $category_image,
+                'category_image' => $this->categoryRepository->getModel()->find($subcategory->category_id)->image_url,
             ]
         );
     }
 
     public function update(Store $request, $id)
     {
-        // $this->authorize('update', [new Subcategory(), $id]);
         $subcategory = $this->subcategoryRepository->getModel()->find($id);
+        $this->authorize('author', $subcategory);
         $data = $request->validated();
 
         if ($request->input('public') != NULL) $data['public'] = true;
         else $data['public'] = false;
-
-        // var_dump($data);
-        // die();
 
         $subcategory->update($data);
         return redirect(url()->previous())
@@ -122,6 +120,7 @@ class SubcategoryController extends Controller
     public function changeVisibility($id)
     {
         $subcategory = $this->subcategoryRepository->getModel()->find($id);
+        $this->authorize('author', $subcategory);
         $subcategory->update(['hidden' => !$subcategory->hidden]);
 
         return redirect(url()->previous())
@@ -131,13 +130,12 @@ class SubcategoryController extends Controller
     //!  DELETE
     public function delete(Request $request, $id)
     {
-        // $this->authorize('delete', [new Subcategory(), $id]);
         $subcategory = $this->subcategoryRepository->getModel()->find($id);
-        $category_id = $subcategory->category_id;
+        $this->authorize('author', $subcategory);
         $subcategory->deleteWithContent($this->pageRepository);
 
         return redirect()
-            ->route('category.show', ['id' => $category_id, 'view' => $request->input('view')])
+            ->route('category.show', ['id' => $subcategory->category_id, 'view' => $request->input('view')])
             ->with('success', 'Podkategoria została usunięta');
     }
 }
